backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
blinker==1.9.0
certifi==2025.4.26
click==8.2.1
distro==1.9.0
Flask==3.1.1
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
MarkupSafe==3.0.2
openai==1.86.0
psycopg2-binary==2.9.10
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.10.1
python-dotenv==1.1.0
sniffio==1.3.1
SQLAlchemy==2.0.41
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
Werkzeug==3.1.3
Flask-Babel>=2.0
flask_jwt_extended>=4.0.0
gunicorn==21.2.0
psutil==5.9.8
flask-talisman==1.1.0
flask-limiter==3.5.0